feature: amount 타입 변경

+++ src/main/java/com/practice/paymentassignment/dto/PaymentPrepareRequest.java
@@ -3,11 +3,13 @@ package com.practice.paymentassignment.dto;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @RequiredArgsConstructor
 public class PaymentPrepareRequest {
     private final Long userId;
     private final Long merchantId;
-    private final int amount;
+    private final BigDecimal amount;
 
 }
